+++ modules/gui/src/MetaObject/params/ui/Qt/Containers.hpp
@@ -62,7 +62,7 @@ namespace mo
                 {
                     if (data.size())
                     {
-                        index->setMaximum(data.size());
+                        index->setMaximum(static_cast<int>(data.size()));
                         if (index->value() < data.size())
                             _data_handler.updateUi(data[index->value()]);
                     }
